requirements.txt
torch
pytest
torchvision
numpy
streamlit
streamlit-webrtc
opencv-python
easyocr
Pillow